change: [UIE-7308] - Notifications for database resize events (#10262)
## Description 📝

**Needs to be tested on DEV/ALPHA**

Bell icon / Alert changes:
System notify users when database resizing starts, display a progress bar which disappears upon completion, update status from 'Active' to 'Resizing', show resizing progress in percentage, and revert to 'Active' status when finished.
Option to switch between dedicated and shared plans, with the exception of switching from dedicated plan to a shared plan of the same size.

## Changes  🔄
- Notification / Event of resizing Start
- Notification progress bar in Bell icon Alert 
- Notification complete – bar disappears 
- Shared to dedicated and back plan change 
- Status Changes from Active to Resizing
- Rendering progress % numbers in UI 
- Changing back to Active when done

## How to test 🧪
**Needs to be tested in Alpha.**
Open application, navigate to /databases, select the database cluster. Go to the Resize tab, choose a plan, and click the 'Resize Database Cluster' button. In the popup window, confirm the resizing by entering the name of the database cluster, then click 'Resize Cluster' button. You can monitor the resizing status in the Summary and Resize tabs, as well as in a table on the Databases page. Additionally, notification about scheduled resizing and a progress bar which disappears upon completion in bell icon dropdown.